+++ packages/client/src/rtc/__tests__/Subscriber.test.ts
@@ -5,6 +5,8 @@ import { Dispatcher } from '../Dispatcher';
 import { StreamSfuClient } from '../../StreamSfuClient';
 import { Subscriber } from '../Subscriber';
 import { CallState } from '../../store';
+import { SfuEvent } from '../../gen/video/sfu/event/events';
+import { PeerType } from '../../gen/video/sfu/models/models';
 
 vi.mock('../../StreamSfuClient', () => {
   console.log('MOCKING StreamSfuClient');
@@ -17,9 +19,10 @@ describe('Subscriber', () => {
   let sfuClient: StreamSfuClient;
   let subscriber: Subscriber;
   let state = new CallState();
+  let dispatcher: Dispatcher;
 
   beforeEach(() => {
-    const dispatcher = new Dispatcher();
+    dispatcher = new Dispatcher();
     sfuClient = new StreamSfuClient({
       dispatcher,
       sfuServer: {
@@ -35,12 +38,14 @@ describe('Subscriber', () => {
       dispatcher,
       state,
       connectionConfig: { iceServers: [] },
+      iceRestartDelay: 100,
     });
   });
 
   afterEach(() => {
     vi.clearAllMocks();
     vi.resetModules();
+    dispatcher.offAll();
   });
 
   describe('Subscriber migration', () => {
@@ -116,6 +121,84 @@ describe('Subscriber', () => {
         onTrack,
       );
       expect(oldPeerConnection.close).toHaveBeenCalled();
+    });
+  });
+
+  describe('Subscriber ICE restart', () => {
+    it('should perform ICE restart when iceRestart event is received', () => {
+      sfuClient.iceRestart = vi.fn();
+      dispatcher.dispatch(
+        SfuEvent.create({
+          eventPayload: {
+            oneofKind: 'iceRestart',
+            iceRestart: {
+              peerType: PeerType.SUBSCRIBER,
+            },
+          },
+        }),
+      );
+
+      expect(sfuClient.iceRestart).toHaveBeenCalledWith({
+        peerType: PeerType.SUBSCRIBER,
+      });
+    });
+
+    it('should not perform ICE restart when iceRestart event is received for a different peer type', () => {
+      sfuClient.iceRestart = vi.fn();
+      dispatcher.dispatch(
+        SfuEvent.create({
+          eventPayload: {
+            oneofKind: 'iceRestart',
+            iceRestart: {
+              peerType: PeerType.PUBLISHER_UNSPECIFIED,
+            },
+          },
+        }),
+      );
+
+      expect(sfuClient.iceRestart).not.toHaveBeenCalled();
+    });
+
+    it(`should drop consequent ICE restart requests`, async () => {
+      sfuClient.iceRestart = vi.fn();
+      // @ts-ignore
+      subscriber['pc'].signalingState = 'have-remote-offer';
+
+      await subscriber.restartIce();
+      expect(sfuClient.iceRestart).not.toHaveBeenCalled();
+    });
+
+    it(`should perform ICE restart when connection state changes to 'failed'`, () => {
+      vi.spyOn(subscriber, 'restartIce').mockResolvedValue();
+      // @ts-ignore
+      subscriber['pc'].iceConnectionState = 'failed';
+      subscriber['onIceConnectionStateChange']();
+      expect(subscriber.restartIce).toHaveBeenCalled();
+    });
+
+    it(`should perform ICE restart when connection state changes to 'disconnected'`, () => {
+      vi.spyOn(subscriber, 'restartIce').mockResolvedValue();
+      vi.useFakeTimers();
+
+      // @ts-ignore
+      subscriber['pc'].iceConnectionState = 'disconnected';
+      subscriber['onIceConnectionStateChange']();
+      vi.runAllTimers();
+      expect(subscriber.restartIce).toHaveBeenCalled();
+    });
+
+    it(`should bail-out from ICE restart once connection recovers before timeout`, () => {
+      vi.spyOn(subscriber, 'restartIce').mockResolvedValue();
+      vi.useFakeTimers();
+
+      // @ts-ignore
+      subscriber['pc'].iceConnectionState = 'disconnected';
+      subscriber['onIceConnectionStateChange']();
+      // @ts-ignore
+      subscriber['pc'].iceConnectionState = 'connected';
+
+      vi.runAllTimers();
+      expect(subscriber.restartIce).not.toHaveBeenCalled();
     });
   });
 });