+++ packages/react-native-sdk/src/icons/PhoneDown.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Path, Svg } from 'react-native-svg';
 import { A11yIcons } from '../constants/A11yLabels';