+++ sample-apps/react-native/dogfood/src/components/VideoWrapper.tsx
@@ -25,7 +25,7 @@ export const VideoWrapper = ({ children }: PropsWithChildren<{}>) => {
     const user = {
       id: userId,
       name: userName,
-      imageUrl: userImageUrl,
+      image: userImageUrl,
     };
     const _videoClient = new StreamVideoClient({
       apiKey: STREAM_API_KEY,
@@ -46,7 +46,12 @@ export const VideoWrapper = ({ children }: PropsWithChildren<{}>) => {
   }
 
   return (
-    <StreamVideo client={videoClient} translationsOverrides={translations}>
+    <StreamVideo
+      client={videoClient}
+      // By default we keep English(en) for now, unless we start supporting switching of language in the application
+      language={'en'}
+      translationsOverrides={translations}
+    >
       {children}
     </StreamVideo>
   );