+++ sample-apps/react-native/dogfood/src/screens/Call/JoinCallScreen.tsx
@@ -10,46 +10,23 @@ import {
   View,
 } from 'react-native';
 import { useAppGlobalStoreValue } from '../../contexts/AppContext';
-import { useStreamVideoClient } from '@stream-io/video-react-native-sdk';
+import {
+  useI18n,
+  useStreamVideoClient,
+} from '@stream-io/video-react-native-sdk';
 import { MemberRequest } from '@stream-io/video-client';
-import { v4 as uuidv4 } from 'uuid';
 import { appTheme } from '../../theme';
 import { Button } from '../../components/Button';
 import { TextInput } from '../../components/TextInput';
+import { KnownUsers } from '../../constants/KnownUsers';
+import { randomId } from '../../modules/helpers/randomId';
 
 const JoinCallScreen = () => {
   const [ringingUserIdsText, setRingingUserIdsText] = useState<string>('');
   const userId = useAppGlobalStoreValue((store) => store.userId);
   const [ringingUsers, setRingingUsers] = useState<string[]>([]);
   const videoClient = useStreamVideoClient();
-
-  const users = [
-    {
-      id: 'steve',
-      name: 'Steve Galilli',
-      image: 'https://ca.slack-edge.com/T02RM6X6B-U039J798FJ7-894cf1f07326-512',
-    },
-    {
-      id: 'khushal',
-      name: 'Khushal Agarwal',
-      image: 'https://ca.slack-edge.com/T02RM6X6B-U02DTREQ2KX-5d600c87d3bc-512',
-    },
-    {
-      id: 'santhosh',
-      name: 'Santhosh Vaiyapuri',
-      image: 'https://ca.slack-edge.com/T02RM6X6B-U0359AX2TUY-dc7dbec0bb88-512',
-    },
-    {
-      id: 'oliver',
-      name: 'Oliver Lazoroski',
-      image: 'https://ca.slack-edge.com/T02RM6X6B-U03HJKTMSQZ-cdf636547793-512',
-    },
-    {
-      id: 'zita',
-      name: 'Zita Szupera',
-      image: 'https://ca.slack-edge.com/T02RM6X6B-U02CA8MV9D1-8631020b96bf-512',
-    },
-  ];
+  const { t } = useI18n();
 
   const startCallHandler = useCallback(async () => {
     let ringingUserIds = !ringingUserIdsText
@@ -60,7 +37,7 @@ const JoinCallScreen = () => {
     ringingUserIds = [...new Set([...ringingUserIds, userId])];
 
     try {
-      const call = videoClient?.call('default', uuidv4().toLowerCase());
+      const call = videoClient?.call('default', randomId());
       await call?.getOrCreate({
         ring: true,
         data: {
@@ -100,44 +77,42 @@ const JoinCallScreen = () => {
       behavior={Platform.OS === 'ios' ? 'position' : 'padding'}
     >
       <View>
-        <Text style={styles.headerText}>Select Participants</Text>
-        {users
-          .filter((user) => user.id !== userId)
-          .map((user) => {
-            return (
-              <Pressable
-                style={styles.participant}
-                key={user.id}
-                onPress={() => ringingUsersSetHandler(user.id)}
+        <Text style={styles.headerText}>{t('Select Participants')}</Text>
+        {KnownUsers.filter((user) => user.id !== userId).map((user) => {
+          return (
+            <Pressable
+              style={styles.participant}
+              key={user.id}
+              onPress={() => ringingUsersSetHandler(user.id)}
+            >
+              <Image source={{ uri: user.image }} style={styles.avatar} />
+              <Text
+                style={[
+                  styles.text,
+                  isRingingUserSelected(user.id)
+                    ? styles.selectedParticipant
+                    : null,
+                ]}
               >
-                <Image source={{ uri: user.image }} style={styles.avatar} />
-                <Text
-                  style={[
-                    styles.text,
-                    isRingingUserSelected(user.id)
-                      ? styles.selectedParticipant
-                      : null,
-                  ]}
-                >
-                  {user.name + ' - id: ' + user.id}
-                </Text>
-              </Pressable>
-            );
-          })}
-        <Text style={styles.orText}>Or</Text>
+                {user.name + ' - id: ' + user.id}
+              </Text>
+            </Pressable>
+          );
+        })}
+        <Text style={styles.orText}>{t('OR')}</Text>
         <TextInput
           autoCapitalize="none"
           autoCorrect={false}
-          placeholder="Enter comma separated User ids"
+          placeholder={t('Enter comma separated User ids')}
           value={ringingUserIdsText}
           onChangeText={(value) => {
             setRingingUserIdsText(value);
           }}
-          textInputStyle={styles.textInputStyle}
+          style={styles.textInputStyle}
         />
       </View>
       <Button
-        title="Start a New Call"
+        title={t('Start a New Call')}
         disabled={ringingUserIdsText === '' && ringingUsers.length === 0}
         onPress={startCallHandler}
       />