+++ sample-apps/react-native/dogfood/src/screens/LoginScreen.tsx
@@ -15,6 +15,8 @@ import { useAppGlobalStoreSetState } from '../contexts/AppContext';
 import { appTheme } from '../theme';
 import { Button } from '../components/Button';
 import { TextInput } from '../components/TextInput';
+import { useI18n } from '@stream-io/video-react-native-sdk';
+import { KnownUsers } from '../constants/KnownUsers';
 
 GoogleSignin.configure({
   // webClientId: '<FROM DEVELOPER CONSOLE>', // client ID of type WEB for your server (needed to verify user ID and offline access)
@@ -35,13 +37,19 @@ const generateValidUserId = (userId: string) => {
 const LoginScreen = () => {
   const [localUserId, setLocalUserId] = useState('');
   const [loader, setLoader] = useState(false);
+  const { t } = useI18n();
 
   const setState = useAppGlobalStoreSetState();
 
   const loginHandler = async () => {
     try {
       const _userId = generateValidUserId(localUserId);
-      const _userImageUrl = `https://getstream.io/random_png/?id=${_userId}&name=${_userId}`;
+      let _userImageUrl = `https://getstream.io/random_png/?id=${_userId}&name=${_userId}`;
+      const _user = KnownUsers.find((u) => u.id === _userId);
+      if (_user) {
+        _userImageUrl = _user.image;
+      }
+
       setState({
         userId: _userId,
         userName: _userId,
@@ -83,16 +91,18 @@ const LoginScreen = () => {
     >
       <Image source={require('../assets/Logo.png')} style={styles.logo} />
       <View>
-        <Text style={styles.title}>Stream DogFood App</Text>
+        <Text style={styles.title}>{t('Stream DogFood App')}</Text>
         <Text style={styles.subTitle}>
-          Please sign in with your Google Stream account or a Custom user id.
+          {t(
+            'Please sign in with your Google Stream account or a Custom user id.',
+          )}
         </Text>
       </View>
 
       <View style={styles.bottomView}>
         <View style={styles.customUser}>
           <TextInput
-            placeholder="Enter custom user"
+            placeholder={t('Enter custom user')}
             value={localUserId}
             onChangeText={(text) => {
               setLocalUserId(text);
@@ -101,15 +111,15 @@ const LoginScreen = () => {
             autoCorrect={false}
           />
           <Button
-            title="Login"
+            title={t('Login')}
             disabled={!localUserId}
             onPress={loginHandler}
             buttonStyle={styles.loginButton}
           />
         </View>
-        <Text style={styles.orText}>OR</Text>
+        <Text style={styles.orText}>{t('OR')}</Text>
         <Button
-          title="Google Sign In"
+          title={t('Google Sign In')}
           onPress={signInViaGoogle}
           disabled={loader}
           buttonStyle={styles.googleSignin}