+++ sample-apps/react/livestream-app/src/hooks/useSetCall.ts
@@ -1,6 +1,10 @@
 import { useParams } from 'react-router-dom';
 import { useEffect, useState } from 'react';
-import { Call, StreamVideoClient } from '@stream-io/video-react-sdk';
+import {
+  Call,
+  CallingState,
+  StreamVideoClient,
+} from '@stream-io/video-react-sdk';
 import { getURLCredentials } from '../utils/getURLCredentials';
 import { DEFAULT_CALL_TYPE } from '../utils/constants';
 
@@ -13,7 +17,15 @@ export const useSetCall = (client?: StreamVideoClient) => {
     if (!(client && callId)) {
       return;
     }
-    setCall(client.call(type ?? DEFAULT_CALL_TYPE, callId));
+    const _call = client.call(type ?? DEFAULT_CALL_TYPE, callId);
+    setCall(_call);
+
+    return () => {
+      if (_call?.state.callingState !== CallingState.LEFT) {
+        _call?.leave();
+      }
+      setCall(undefined);
+    };
   }, [client, callId, type]);
 
   return call;