+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,152 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1-alpha.342](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.341...@stream-io/video-react-native-sdk-0.0.1-alpha.342) (2023-07-27)
+
+
+
+## [0.0.1-alpha.341](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.340...@stream-io/video-react-native-sdk-0.0.1-alpha.341) (2023-07-26)
+
+
+
+## [0.0.1-alpha.340](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.339...@stream-io/video-react-native-sdk-0.0.1-alpha.340) (2023-07-26)
+
+
+
+## [0.0.1-alpha.339](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.338...@stream-io/video-react-native-sdk-0.0.1-alpha.339) (2023-07-26)
+
+
+
+## [0.0.1-alpha.338](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.337...@stream-io/video-react-native-sdk-0.0.1-alpha.338) (2023-07-26)
+
+
+
+## [0.0.1-alpha.337](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.336...@stream-io/video-react-native-sdk-0.0.1-alpha.337) (2023-07-26)
+
+
+### Bug Fixes
+
+* **react-native:** broken tests due to translations ([#855](https://github.com/GetStream/stream-video-js/issues/855)) ([63646fe](https://github.com/GetStream/stream-video-js/commit/63646fee0672d334f72c674013d31603bc910f4c))
+
+
+### Features
+
+* **react-native:** avatar docs ([#853](https://github.com/GetStream/stream-video-js/issues/853)) ([6ef7f62](https://github.com/GetStream/stream-video-js/commit/6ef7f62aaea8d4f3a8b5318c2f60fd2c0ba5d598))
+
+
+
+## [0.0.1-alpha.336](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.335...@stream-io/video-react-native-sdk-0.0.1-alpha.336) (2023-07-25)
+
+
+### Features
+
+* **react-native:** add translations to SDK and DF app ([#828](https://github.com/GetStream/stream-video-js/issues/828)) ([c7a7f73](https://github.com/GetStream/stream-video-js/commit/c7a7f73b5cfd9222101e4c44b6c9ec42006bcac2))
+
+
+
+## [0.0.1-alpha.335](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.334...@stream-io/video-react-native-sdk-0.0.1-alpha.335) (2023-07-25)
+
+
+
+## [0.0.1-alpha.334](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.333...@stream-io/video-react-native-sdk-0.0.1-alpha.334) (2023-07-25)
+
+
+
+## [0.0.1-alpha.333](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.332...@stream-io/video-react-native-sdk-0.0.1-alpha.333) (2023-07-25)
+
+
+
+## [0.0.1-alpha.332](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.331...@stream-io/video-react-native-sdk-0.0.1-alpha.332) (2023-07-25)
+
+
+### Bug Fixes
+
+* **react-native:** wrong import in audio room ([14f4e6b](https://github.com/GetStream/stream-video-js/commit/14f4e6bfa3547849a9a7555d4fae50405ffb1a3f))
+
+
+
+## [0.0.1-alpha.331](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.330...@stream-io/video-react-native-sdk-0.0.1-alpha.331) (2023-07-25)
+
+
+
+## [0.0.1-alpha.330](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.329...@stream-io/video-react-native-sdk-0.0.1-alpha.330) (2023-07-24)
+
+
+
+## [0.0.1-alpha.329](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.328...@stream-io/video-react-native-sdk-0.0.1-alpha.329) (2023-07-24)
+
+
+
+## [0.0.1-alpha.328](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.327...@stream-io/video-react-native-sdk-0.0.1-alpha.328) (2023-07-24)
+
+
+
+## [0.0.1-alpha.327](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.326...@stream-io/video-react-native-sdk-0.0.1-alpha.327) (2023-07-22)
+
+
+
+## [0.0.1-alpha.326](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.325...@stream-io/video-react-native-sdk-0.0.1-alpha.326) (2023-07-21)
+
+
+### Features
+
+* **react-native:** add permissions to docs ([#842](https://github.com/GetStream/stream-video-js/issues/842)) ([1ec9c4d](https://github.com/GetStream/stream-video-js/commit/1ec9c4d3e2b48c35d6dfcf6efd4460aac8a1e9e5))
+
+
+
+## [0.0.1-alpha.325](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.324...@stream-io/video-react-native-sdk-0.0.1-alpha.325) (2023-07-21)
+
+
+### Features
+
+* **react-native:** add camera mic documentation ([#834](https://github.com/GetStream/stream-video-js/issues/834)) ([ffd564f](https://github.com/GetStream/stream-video-js/commit/ffd564f18b0d3179580fc4e3a6cb308220a6ca53))
+
+
+
+## [0.0.1-alpha.324](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.323...@stream-io/video-react-native-sdk-0.0.1-alpha.324) (2023-07-21)
+
+
+### Bug Fixes
+
+* strict mode issue ([#740](https://github.com/GetStream/stream-video-js/issues/740)) ([c39e4e4](https://github.com/GetStream/stream-video-js/commit/c39e4e4041a2326393478ad808b2aa791d50f8ce))
+
+
+
+## [0.0.1-alpha.323](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.322...@stream-io/video-react-native-sdk-0.0.1-alpha.323) (2023-07-21)
+
+
+
+## [0.0.1-alpha.322](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.321...@stream-io/video-react-native-sdk-0.0.1-alpha.322) (2023-07-21)
+
+
+
+## [0.0.1-alpha.321](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.320...@stream-io/video-react-native-sdk-0.0.1-alpha.321) (2023-07-20)
+
+
+### Features
+
+* **react-native:** add audio rooms in dogfood and related docs ([#810](https://github.com/GetStream/stream-video-js/issues/810)) ([2346d8f](https://github.com/GetStream/stream-video-js/commit/2346d8f265bec51f7e9425fcec266aefe033405d))
+
+
+
+## [0.0.1-alpha.320](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.319...@stream-io/video-react-native-sdk-0.0.1-alpha.320) (2023-07-20)
+
+
+
+## [0.0.1-alpha.319](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.318...@stream-io/video-react-native-sdk-0.0.1-alpha.319) (2023-07-20)
+
+
+### Bug Fixes
+
+* server side user connect + add tests ([#825](https://github.com/GetStream/stream-video-js/issues/825)) ([95ea24d](https://github.com/GetStream/stream-video-js/commit/95ea24d03306d1b25c3c5af042a202a7b551d865))
+
+
+### Features
+
+* **react-native:** add docs for keeping call alive ([#817](https://github.com/GetStream/stream-video-js/issues/817)) ([1da5c38](https://github.com/GetStream/stream-video-js/commit/1da5c385fad8529bba5ac407915a3b09fbaaceec))
+
+
+
 ## [0.0.1-alpha.318](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.317...@stream-io/video-react-native-sdk-0.0.1-alpha.318) (2023-07-19)